+++ packages/contracts-nest/src/user/user.services.ts
@@ -5,10 +5,12 @@
 // source: volontariapp/user/user.services.proto
 
 /* eslint-disable */
-import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
-import { Observable } from 'rxjs';
+import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
+import { Observable } from "rxjs";
 import {
   AddBadgeToUserCommand,
+  AdminDeleteUserCommand,
+  AdminUpdateUserCommand,
   CreateBadgeCommand,
   DeleteBadgeCommand,
   DeleteUserCommand,
@@ -19,16 +21,20 @@ import {
   SignUpCommand,
   UpdateBadgeCommand,
   UpdateUserCommand,
-} from './user.command.js';
+} from "./user.command.js";
 import {
+  AdminGetUserQuery,
   GetBadgeBySlugQuery,
   GetBadgeQuery,
   GetUserQuery,
   ListBadgesQuery,
   ListUsersQuery,
-} from './user.query.js';
+} from "./user.query.js";
 import {
   AddBadgeToUserResponse,
+  AdminDeleteUserResponse,
+  AdminUpdateUserResponse,
+  AdminUserResponse,
   BadgeResponse,
   CreateBadgeResponse,
   DeleteBadgeResponse,
@@ -44,7 +50,7 @@ import {
   UpdateBadgeResponse,
   UpdateUserResponse,
   UserResponse,
-} from './user.responses.js';
+} from "./user.responses.js";
 
 export interface UserServiceClient {
   getUser(request: GetUserQuery): Observable<UserResponse>;
@@ -57,13 +63,17 @@ export interface UserServiceClient {
 
   deleteUser(request: DeleteUserCommand): Observable<DeleteUserResponse>;
 
+  adminGetUser(request: AdminGetUserQuery): Observable<AdminUserResponse>;
+
+  adminUpdateUser(request: AdminUpdateUserCommand): Observable<AdminUpdateUserResponse>;
+
+  adminDeleteUser(request: AdminDeleteUserCommand): Observable<AdminDeleteUserResponse>;
+
   login(request: LoginCommand): Observable<LoginResponse>;
 
   refreshToken(request: RefreshTokenCommand): Observable<RefreshTokenResponse>;
 
-  incrementImpactScore(
-    request: IncrementImpactScoreCommand,
-  ): Observable<IncrementImpactScoreResponse>;
+  incrementImpactScore(request: IncrementImpactScoreCommand): Observable<IncrementImpactScoreResponse>;
 
   addBadgeToUser(request: AddBadgeToUserCommand): Observable<AddBadgeToUserResponse>;
 
@@ -73,13 +83,9 @@ export interface UserServiceClient {
 export interface UserServiceController {
   getUser(request: GetUserQuery): Promise<UserResponse> | Observable<UserResponse> | UserResponse;
 
-  listUsers(
-    request: ListUsersQuery,
-  ): Promise<ListUsersResponse> | Observable<ListUsersResponse> | ListUsersResponse;
+  listUsers(request: ListUsersQuery): Promise<ListUsersResponse> | Observable<ListUsersResponse> | ListUsersResponse;
 
-  signUp(
-    request: SignUpCommand,
-  ): Promise<SignUpResponse> | Observable<SignUpResponse> | SignUpResponse;
+  signUp(request: SignUpCommand): Promise<SignUpResponse> | Observable<SignUpResponse> | SignUpResponse;
 
   updateUser(
     request: UpdateUserCommand,
@@ -89,6 +95,18 @@ export interface UserServiceController {
     request: DeleteUserCommand,
   ): Promise<DeleteUserResponse> | Observable<DeleteUserResponse> | DeleteUserResponse;
 
+  adminGetUser(
+    request: AdminGetUserQuery,
+  ): Promise<AdminUserResponse> | Observable<AdminUserResponse> | AdminUserResponse;
+
+  adminUpdateUser(
+    request: AdminUpdateUserCommand,
+  ): Promise<AdminUpdateUserResponse> | Observable<AdminUpdateUserResponse> | AdminUpdateUserResponse;
+
+  adminDeleteUser(
+    request: AdminDeleteUserCommand,
+  ): Promise<AdminDeleteUserResponse> | Observable<AdminDeleteUserResponse> | AdminDeleteUserResponse;
+
   login(request: LoginCommand): Promise<LoginResponse> | Observable<LoginResponse> | LoginResponse;
 
   refreshToken(
@@ -97,10 +115,7 @@ export interface UserServiceController {
 
   incrementImpactScore(
     request: IncrementImpactScoreCommand,
-  ):
-    | Promise<IncrementImpactScoreResponse>
-    | Observable<IncrementImpactScoreResponse>
-    | IncrementImpactScoreResponse;
+  ): Promise<IncrementImpactScoreResponse> | Observable<IncrementImpactScoreResponse> | IncrementImpactScoreResponse;
 
   addBadgeToUser(
     request: AddBadgeToUserCommand,
@@ -108,39 +123,39 @@ export interface UserServiceController {
 
   removeBadgeFromUser(
     request: RemoveBadgeFromUserCommand,
-  ):
-    | Promise<RemoveBadgeFromUserResponse>
-    | Observable<RemoveBadgeFromUserResponse>
-    | RemoveBadgeFromUserResponse;
+  ): Promise<RemoveBadgeFromUserResponse> | Observable<RemoveBadgeFromUserResponse> | RemoveBadgeFromUserResponse;
 }
 
 export function UserServiceControllerMethods() {
   return function (constructor: Function) {
     const grpcMethods: string[] = [
-      'getUser',
-      'listUsers',
-      'signUp',
-      'updateUser',
-      'deleteUser',
-      'login',
-      'refreshToken',
-      'incrementImpactScore',
-      'addBadgeToUser',
-      'removeBadgeFromUser',
+      "getUser",
+      "listUsers",
+      "signUp",
+      "updateUser",
+      "deleteUser",
+      "adminGetUser",
+      "adminUpdateUser",
+      "adminDeleteUser",
+      "login",
+      "refreshToken",
+      "incrementImpactScore",
+      "addBadgeToUser",
+      "removeBadgeFromUser",
     ];
     for (const method of grpcMethods) {
       const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
-      GrpcMethod('UserService', method)(constructor.prototype[method], method, descriptor);
+      GrpcMethod("UserService", method)(constructor.prototype[method], method, descriptor);
     }
     const grpcStreamMethods: string[] = [];
     for (const method of grpcStreamMethods) {
       const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
-      GrpcStreamMethod('UserService', method)(constructor.prototype[method], method, descriptor);
+      GrpcStreamMethod("UserService", method)(constructor.prototype[method], method, descriptor);
     }
   };
 }
 
-export const USER_SERVICE_NAME = 'UserService';
+export const USER_SERVICE_NAME = "UserService";
 
 export interface BadgeServiceClient {
   createBadge(request: CreateBadgeCommand): Observable<CreateBadgeResponse>;
@@ -169,9 +184,7 @@ export interface BadgeServiceController {
     request: DeleteBadgeCommand,
   ): Promise<DeleteBadgeResponse> | Observable<DeleteBadgeResponse> | DeleteBadgeResponse;
 
-  getBadge(
-    request: GetBadgeQuery,
-  ): Promise<BadgeResponse> | Observable<BadgeResponse> | BadgeResponse;
+  getBadge(request: GetBadgeQuery): Promise<BadgeResponse> | Observable<BadgeResponse> | BadgeResponse;
 
   listBadges(
     request: ListBadgesQuery,
@@ -185,23 +198,23 @@ export interface BadgeServiceController {
 export function BadgeServiceControllerMethods() {
   return function (constructor: Function) {
     const grpcMethods: string[] = [
-      'createBadge',
-      'updateBadge',
-      'deleteBadge',
-      'getBadge',
-      'listBadges',
-      'getBadgeBySlug',
+      "createBadge",
+      "updateBadge",
+      "deleteBadge",
+      "getBadge",
+      "listBadges",
+      "getBadgeBySlug",
     ];
     for (const method of grpcMethods) {
       const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
-      GrpcMethod('BadgeService', method)(constructor.prototype[method], method, descriptor);
+      GrpcMethod("BadgeService", method)(constructor.prototype[method], method, descriptor);
     }
     const grpcStreamMethods: string[] = [];
     for (const method of grpcStreamMethods) {
       const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
-      GrpcStreamMethod('BadgeService', method)(constructor.prototype[method], method, descriptor);
+      GrpcStreamMethod("BadgeService", method)(constructor.prototype[method], method, descriptor);
     }
   };
 }
 
-export const BADGE_SERVICE_NAME = 'BadgeService';
+export const BADGE_SERVICE_NAME = "BadgeService";